+++ src/screens/login.js
@@ -1,39 +1,38 @@
 import React from 'react';
 import { Text, View, TouchableOpacity, TextInput, Alert } from 'react-native';
-import loginStyles from '../styles/LoginStyles';
 
 function login({ navigation }) {
   return (
-    <View style={loginStyles.container}>
-      <Text style={loginStyles.txtTitle}>
+    <View style={styles.container}>
+      <Text style={styles.txtTitle}>
         Log In
         </Text>
       <TouchableOpacity onPress={() => navigation.navigate('ChooseAcc')}>
-        <Text style={loginStyles.txtCreate}>
-          <Text style={loginStyles.txtNewUser}>
+        <Text style={styles.txtCreate}>
+          <Text style={styles.txtNewUser}>
             New user? </Text> Create an account
         </Text>
       </TouchableOpacity>
-      <View style={loginStyles.viewInput}>
-        <TextInput style={loginStyles.txtInput}
+      <View style={styles.viewInput}>
+        <TextInput style={styles.txtInput}
           placeholder='Email'
           placeholderTextColor='black' />
       </View>
-      <View style={loginStyles.viewInput}>
-        <TextInput style={loginStyles.txtInput}
+      <View style={styles.viewInput}>
+        <TextInput style={styles.txtInput}
           secureTextEntry
           placeholder='Password'
           placeholderTextColor='black' />
       </View>
       <TouchableOpacity
         onPress={() => Alert.alert('FORGOT!')}>
-        <Text style={loginStyles.txtForgot}>
+        <Text style={styles.txtForgot}>
           Forgot Password?
         </Text>
       </TouchableOpacity>
-      <TouchableOpacity style={loginStyles.btnLogin}
+      <TouchableOpacity style={styles.btnLogin}
         onPress={() => Alert.alert('LOGIN!')}>
-        <Text style={loginStyles.btnTxt}>
+        <Text style={styles.btnTxt}>
           Log In
         </Text>
       </TouchableOpacity>
@@ -42,3 +41,69 @@ function login({ navigation }) {
 }
 
 export default login;
+
+
+const styles = ({
+  container: {
+      alignItems: 'center',
+      justifyContent: 'center',
+  },
+  txtTitle: {
+    fontSize: 40,
+    textAlign: 'center',
+    color: 'black',
+    fontWeight: 'bold',
+    marginTop: '25%',
+  },
+  txtNewUser: {
+      fontSize: 15,
+      color: 'black',
+      letterSpacing: -1,
+  },
+  txtCreate: {
+      fontSize: 15,
+      color: '#77A8AB',
+      letterSpacing: -1,
+  },
+  viewInput: {
+    fontSize: 18,
+    backgroundColor: 'white',
+    borderRadius: 25,
+    height: 50,
+    width: 300,
+    padding: 10,
+    marginTop: 15,
+    justifyContent:'center',
+  },
+  txtInput: {
+      height:50,
+  },
+  txtForgot: {
+      fontSize: 15,
+      color: '#77A8AB',
+      marginTop: '5%',
+  },
+  btnLogin: {
+      marginTop: '5%',
+      borderRadius: 20,
+      paddingVertical: 10,
+      paddingHorizontal: 12,
+      backgroundColor: '#77A8AB',
+      width: 170,
+      height: 45,
+      shadowColor: "#000",
+      shadowOffset: {
+          width: 0,
+          height: 2,
+      },
+      shadowOpacity: 0.25,
+      shadowRadius: 3.84,
+      elevation: 5,
+  },
+  btnTxt: {
+      fontSize: 18,
+      fontWeight: 'bold',
+      color: 'black',
+      alignSelf: 'center',
+    }
+})
