--- src/styles/chooseCreateStyles.js
@@ -1,61 +1,0 @@
-const chooseStyles = ({
-    container: {
-        flex: 1,
-        alignItems: 'center',
-    },
-    txtTitle: {
-        fontSize: 40,
-        color: 'black',
-        fontWeight: 'bold',
-        marginTop: '25%',
-        alignItems: 'center',
-    },
-    txtAlready: {
-        fontSize: 15,
-        color: 'black',
-        letterSpacing: -1,
-    },
-    txtSign: {
-        fontSize: 15,
-        color: '#77A8AB',
-        letterSpacing: -1,
-    },
-    viewInput: {
-        fontSize: 18,
-        backgroundColor: 'white',
-        borderRadius: 25,
-        height: 50,
-        width: 300,
-        padding: 10,
-        marginTop: 15,
-        justifyContent: 'center',
-    },
-    txtInput: {
-        height: 50,
-    },
-    btn: {
-        marginTop: '5%',
-        borderRadius: 20,
-        paddingVertical: 10,
-        paddingHorizontal: 12,
-        backgroundColor: '#77A8AB',
-        width: 250,
-        height: 45,
-        shadowColor: "#000",
-        shadowOffset: {
-            width: 0,
-            height: 2,
-        },
-        shadowOpacity: 0.25,
-        shadowRadius: 3.84,
-        elevation: 5,
-    },
-    txtBtn: {
-        fontSize: 18,
-        fontWeight: 'bold',
-        color: 'black',
-        alignSelf: 'center',
-    },
-})
-
-export default chooseStyles
